pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "malak"
version = "0.1.0"
description = "A Python toolkit for edge AI model optimization"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Mohammed Alnemari", email = "mohammedalnemari@gmail.com"},
]
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.optional-dependencies]
onnx = ["onnx>=1.14.0", "onnxruntime>=1.15.0"]
yaml = ["pyyaml>=6.0"]
dev = ["pytest>=7.0", "onnx>=1.14.0", "pyyaml>=6.0"]

[project.scripts]
edgeai = "malak.cli:main"

[project.urls]
Repository = "https://github.com/alnemari-m/malak-platform"

[tool.setuptools.packages.find]
include = ["malak*"]
